add crud policies for bookmark and tags

+++ app/Http/Controllers/Bookmark/DestroyBookmarkController.php
@@ -6,11 +6,16 @@ namespace App\Http\Controllers\Bookmark;
 
 use App\Models\Bookmark;
 use Illuminate\Http\Response;
+use Illuminate\Support\Facades\Auth;
 
 final class DestroyBookmarkController
 {
     public function __invoke(Bookmark $bookmark): Response
     {
+        if (Auth::user()->cannot('delete', $bookmark)) {
+            abort(403);
+        }
+
         $bookmark->delete();
 
         return htmx()->trigger('loadBookmarks')->response();
