+++ app/Http/Controllers/Tag/DeleteTagController.php
@@ -5,11 +5,16 @@ declare(strict_types=1);
 namespace App\Http\Controllers\Tag;
 
 use App\Models\Tag;
+use Illuminate\Support\Facades\Auth;
 
 final class DeleteTagController
 {
     public function __invoke(Tag $tag)
     {
+        if (Auth::user()->cannot('delete', $tag)) {
+            abort(403);
+        }
+
         $tag->bookmarks()->detach();
         $tag->delete();
